loggers

+++ router/routes/productos.js
@@ -4,8 +4,12 @@ const session = require("express-session");
 const passport = require("passport")
 const LocalStrategy = require("passport-local").Strategy
 const numCPUs = require("os").cpus().length
+const compression = require("compression")
 
 const Usuarios = require("../user.js")
+
+const logger = require("../logger.js")
+
 
 let nom = ""
 
@@ -63,38 +67,53 @@ const { json } = require("express");
 const contenedor1 = new contenedor("productos")
 
 router.get('/registro', (req, res) => {
+    logger.info("Esta url existe")
     res.render("registro")
  })
 
 router.get('/login', (req, res) => {
+    logger.info("Esta url existe")
     res.render("login")
  })
 
 router.get('/formulario',(req, res) => {
+    logger.info("Esta url existe")
     res.render("formulario", {name:nom})
  })
 
 router.get('/logout', (req, res) => {
+    logger.info("Esta url existe")
     res.render("logout")
  })
 
 router.get('/productos', (req, res) => {
+    logger.info("Esta url existe")
     contenedor1.getAll(res)
  })
 
  router.get('/productos/:id', (req, res) => {
-    contenedor1.getById(req.params.id, res)
+    logger.info("Esta url existe")
+    if(!isNaN(req.params.id)){
+        logger.info("Parametro ingresado correcto")
+        contenedor1.getById(req.params.id, res)
+    }else{
+        logger.error("Parametro ingresado incorrecto")
+        res.send("Parametro ingresado incorrecto")
+    }
 })
 
 router.get('/errorRegistro', (req, res) => {
+    logger.info("Esta url existe")
     res.render("error2")
  })
 
 router.get('/errorLogin', (req, res) => {
+    logger.info("Esta url existe")
     res.render("error")
  })
 
  router.get("/logout_", (req,res) => {
+    logger.info("Esta url existe")
     req.session.destroy(err => {
         if (err){
        res.json({status: "Logout ERROR", body:err})
@@ -106,10 +125,12 @@ router.get('/errorLogin', (req, res) => {
 })
 
 router.post("/productos", (req, res) => {
+    logger.info("Esta url existe")
     contenedor1.save(req.body).then( ()=> res.send("Producto añadido"))
 })
 
 router.post("/productos-test", (req,res)=> {
+    logger.info("Esta url existe")
     for(let i=0; i < 5; i++) {
         const obj = { name: faker.commerce.product(),
         price: faker.commerce.price(),
@@ -131,6 +152,7 @@ router.post("/login", passport.authenticate("login", {
 }))
 
 router.put("/productos/:id", (req,res)=> {
+    logger.info("Esta url existe")
     knex("users")
     .where({id: req.params.id})
     .update({name: req.body.name, price: req.body.price})
@@ -144,18 +166,40 @@ router.put("/productos/:id", (req,res)=> {
 
 
 router.delete("/productos/:id", (req,res)=> {
-    contenedor1.deleteById(req.params.id, res)
+    logger.info("Esta url existe")
+    if(!isNaN(req.params.id)){
+        logger.info("Parametro ingresado correcto")
+        contenedor1.deleteById(req.params.id, res)
+    }else{
+        logger.error("Parametro ingresado incorrecto")
+        res.send("Parametro ingresado incorrecto")
+    }
 })
 
 router.delete("/productos", (req,res)=> {
+    logger.info("Esta url existe")
     contenedor1.deleteAll(res)
 })
 
 router.get("/info", (req,res)=> {
+    logger.info("Esta url existe")
+    const info = {ArgumentosEntradas: process.argv, SistemaOperativo: process.platform, VersionNode: process.version, Memoria: process.memoryUsage(), Path: process.execPath, ProcessId: process.pid, CarpetaProyecto: process.cwd(),
+                  Procesadores: numCPUs
+    }
+
+    //Prueba con console log
+    //console.log(info)
+    res.send(info)
+})
+
+router.get("/infoZip", compression(), (req,res)=> {
+    logger.info("Esta url existe")
     const info = {ArgumentosEntradas: process.argv, SistemaOperativo: process.platform, VersionNode: process.version, Memoria: process.memoryUsage(), Path: process.execPath, ProcessId: process.pid, CarpetaProyecto: process.cwd(),
                   Procesadores: numCPUs
     }
     res.send(info)
 })
+
+
 
 module.exports = router;
